+++ lexnlp/extract/common/annotations/copyright_annotation.py
@@ -1,7 +1,7 @@
 __author__ = "ContraxSuite, LLC; LexPredict, LLC"
 __copyright__ = "Copyright 2015-2021, ContraxSuite, LLC"
-__license__ = "https://github.com/LexPredict/lexpredict-lexnlp/blob/2.1.0/LICENSE"
-__version__ = "2.1.0"
+__license__ = "https://github.com/LexPredict/lexpredict-lexnlp/blob/2.2.0/LICENSE"
+__version__ = "2.2.0"
 __maintainer__ = "LexPredict, LLC"
 __email__ = "support@contraxsuite.com"
 
@@ -11,11 +11,12 @@ from lexnlp.extract.common.annotations.text_annotation import TextAnnotation
 
 
 class CopyrightAnnotation(TextAnnotation):
-    record_type = 'copyright'
     """
     create an object of CopyrightAnnotation like
     cp = CopyrightAnnotation(name='name', coords=(0, 100), text='text text')
     """
+    record_type = 'copyright'
+
     def __init__(self,
                  coords: Tuple[int, int],
                  locale: str = 'en',