+++ lexnlp/extract/common/annotations/cusip_annotation.py
@@ -1,7 +1,7 @@
 __author__ = "ContraxSuite, LLC; LexPredict, LLC"
 __copyright__ = "Copyright 2015-2021, ContraxSuite, LLC"
-__license__ = "https://github.com/LexPredict/lexpredict-lexnlp/blob/2.1.0/LICENSE"
-__version__ = "2.1.0"
+__license__ = "https://github.com/LexPredict/lexpredict-lexnlp/blob/2.2.0/LICENSE"
+__version__ = "2.2.0"
 __maintainer__ = "LexPredict, LLC"
 __email__ = "support@contraxsuite.com"
 
@@ -11,11 +11,12 @@ from lexnlp.utils.map import Map
 
 
 class CusipAnnotation(TextAnnotation):
-    record_type = 'cusip'
     """
     create an object of CusipAnnotation like
     cp = CusipAnnotation(coords=(0, 100))
     """
+    record_type = 'cusip'
+
     def __init__(self,
                  coords: Tuple[int, int],
                  locale: str = 'en',