+++ lexnlp/extract/common/annotations/duration_annotation.py
@@ -1,7 +1,7 @@
 __author__ = "ContraxSuite, LLC; LexPredict, LLC"
 __copyright__ = "Copyright 2015-2021, ContraxSuite, LLC"
-__license__ = "https://github.com/LexPredict/lexpredict-lexnlp/blob/2.1.0/LICENSE"
-__version__ = "2.1.0"
+__license__ = "https://github.com/LexPredict/lexpredict-lexnlp/blob/2.2.0/LICENSE"
+__version__ = "2.2.0"
 __maintainer__ = "LexPredict, LLC"
 __email__ = "support@contraxsuite.com"
 
@@ -11,11 +11,12 @@ from lexnlp.extract.common.annotations.text_annotation import TextAnnotation
 
 
 class DurationAnnotation(TextAnnotation):
-    record_type = 'duration'
     """
     create an object of DurationAnnotation like
     cp = DurationAnnotation(coords=(0, 100), value='101 ms')
     """
+    record_type = 'duration'
+
     def __init__(
         self,
         coords: Tuple[int, int],
@@ -26,7 +27,8 @@ class DurationAnnotation(TextAnnotation):
         duration_days: Decimal = None,
         duration_type: str = None,
         duration_type_en: str = None,
-        is_complex: bool = False
+        is_complex: bool = False,
+        value_dict: dict = None
     ) -> None:
         super().__init__(
             name='',
@@ -40,6 +42,7 @@ class DurationAnnotation(TextAnnotation):
         self.duration_type: str = duration_type
         self.duration_type_en: str = duration_type_en
         self.is_complex: bool = is_complex
+        self.value_dict: dict = value_dict
 
     def get_cite_value_parts(self) -> List[str]:
         parts = [str(self.amount or ''),