+++ lexnlp/extract/common/annotations/percent_annotation.py
@@ -1,7 +1,7 @@
 __author__ = "ContraxSuite, LLC; LexPredict, LLC"
 __copyright__ = "Copyright 2015-2021, ContraxSuite, LLC"
-__license__ = "https://github.com/LexPredict/lexpredict-lexnlp/blob/2.1.0/LICENSE"
-__version__ = "2.1.0"
+__license__ = "https://github.com/LexPredict/lexpredict-lexnlp/blob/2.2.0/LICENSE"
+__version__ = "2.2.0"
 __maintainer__ = "LexPredict, LLC"
 __email__ = "support@contraxsuite.com"
 
@@ -12,11 +12,12 @@ from lexnlp.utils.map import Map
 
 
 class PercentAnnotation(TextAnnotation):
-    record_type = 'percent'
     """
     create an object of PercentAnnotation like
     cp = PercentAnnotation(coords=(0, 100), value='10 000 USD')
     """
+    record_type = 'percent'
+
     def __init__(
         self,
         coords: Tuple[int, int],