+++ lexnlp/extract/en/durations.py
@@ -4,8 +4,8 @@ This module implements duration extraction functionality in English.
 
 __author__ = "ContraxSuite, LLC; LexPredict, LLC"
 __copyright__ = "Copyright 2015-2021, ContraxSuite, LLC"
-__license__ = "https://github.com/LexPredict/lexpredict-lexnlp/blob/2.1.0/LICENSE"
-__version__ = "2.1.0"
+__license__ = "https://github.com/LexPredict/lexpredict-lexnlp/blob/2.2.0/LICENSE"
+__version__ = "2.2.0"
 __maintainer__ = "LexPredict, LLC"
 __email__ = "support@contraxsuite.com"
 
@@ -77,7 +77,8 @@ class EnDurationParser(DurationParser):
                 amount=amount,
                 duration_type=duration_type,
                 duration_days=duration_days,
-                text=source_text.strip()
+                text=source_text.strip(),
+                value_dict={duration_type: float(amount)}
             )
             all_annotations.append(ant)
         return all_annotations